Module antsee::color

Antsee is a terminal output library which doesn’t output anything to the terminal. It’s designed for writing highly customizeable and user friendly configuration formats

Re-exports

pub use ansi16::Ansi16;

pub use ansi256::Ansi256;

pub use css::CssColors;

pub use rgb::Rgb;

pub use xterm::XtermColors;

Modules

ansi16

ansi16 provides basic terminal colors

ansi256

This module defines a representation of ANSI256 colors

css

This module serves to provide color names for RGB using css color names

rgb

xterm

This module provides color names for the ANSI256 color palette to be used by RGB and ANSI256 colors

Enums

Color